+++ CHANGELOG.md
@@ -8,8 +8,12 @@ SOUP version numbering generally follows the _major.minor.patch_ scheme.  Odd mi
 
 **New fixes in 1.57.x:**
 
-* [mse331640: Syntax highlight and MathJax is not rendered after reloading an edited post](https://meta.stackexchange.com/q/33164)
+* [mse331640: Syntax highlight and MathJax is not rendered after reloading an edited post](https://meta.stackexchange.com/q/331640)
 * [music396: jTab needs to render in Markdown preview](https://music.meta.stackexchange.com/q/396) (music.SE only)
+
+**Removed fixes:**
+
+* [mse322619: Can't see other pages of answers on a certain deleted question](https://meta.stackexchange.com/q/322619)
 
 **Other changes:**